move duplicate code to superclass
