+++ src/sync/core/events/sensor/sensor-updated.event.ts
@@ -17,7 +17,7 @@ export class SensorUpdated extends SensorEvent {
   readonly manufacturer: string;
 
   @Expose()
-  readonly observationArea: object;
+  readonly observationArea: Record<string, any>;
 
   @Expose()
   readonly documentationUrl: string;
@@ -29,12 +29,12 @@ export class SensorUpdated extends SensorEvent {
   readonly typeName: string;
 
   @Expose()
-  readonly typeDetails: object;
+  readonly typeDetails: Record<string, any>;
 
   constructor(sensorId: string, name: string, aim: string, description: string, manufacturer: string,
-              observationArea: object,
+              observationArea: Record<string, any>,
               documentationUrl: string, theme: string[], typeName: string,
-              typeDetails: object) {
+              typeDetails: Record<string, any>) {
     super(sensorId);
     this.name = name;
     this.aim = aim;
